ProyectoBases Version Postgresql/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
SQLAlchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
requests==2.31.0
pytz==2023.3
